+++ src/battle/engine/action_effects.h
@@ -19,10 +19,10 @@ typedef struct {
 } EffectContext;
 #pragma pack(pop)
 
-typedef u16(*PFNSRPUSHEFFECT)(PFNSREFFECTCALLBACK);
-u16 srPushEffect60(PFNSREFFECTCALLBACK callback);
-u16 srPushEffect100(PFNSREFFECTCALLBACK callback);
-u16 srPushEffect10(PFNSREFFECTCALLBACK callback);
+typedef u8(*PFNSRPUSHEFFECT)(PFNSREFFECTCALLBACK);
+u8 srPushEffect60(PFNSREFFECTCALLBACK callback);
+u8 srPushEffect100(PFNSREFFECTCALLBACK callback);
+u8 srPushEffect10(PFNSREFFECTCALLBACK callback);
 EffectContext* getEffectContext60(u16 effectIndex);
 EffectContext* getEffectContext100(u16 effectIndex);
 EffectContext* getEffectContext10(u16 effectIndex);
