+++ src/battle/engine/animation_player.h
@@ -7,8 +7,8 @@
 
 #define GAME_ANGLE_MAX 4096
 
-void srPlayModelAnimation(u16 actorID);
-void srHandleAnimateModel(u16 actorID);
+void srPlayModelAnimation(u8 actorID);
+void srHandleAnimateModel(u8 actorID);
 
 AnimationEvent* getAnimationEvent(u8 queueIndex);
 AnimationEvent* getAnimationEventTop();
