+++ src/battle/engine/animation_script_parser.h
@@ -5,6 +5,6 @@
 #include "animation_script_opcode.h"
 #include "animation_player.h"
 
-void runAnimationScript(u16 actorID, u8** ptrToScriptTable);
+void runAnimationScript(u8 actorID, u8** ptrToScriptTable);
 
 #endif // !ANIMATION_SCRIPT_PARSER_H
